+++ jac/jaclang/runtimelib/machine.py
@@ -1340,6 +1340,15 @@ class JacBasics:
         setattr(func, "__jac_exit", None)  # noqa:B010
         return func
 
+
+class JacByLLM:
+    """Jac byLLM integration."""
+
+    @staticmethod
+    def get_mtir(caller: Callable, args: dict, call_params: dict) -> object:
+        """Get byLLM library."""
+        return None
+
     @staticmethod
     def sem(semstr: str, inner_semstr: dict[str, str]) -> Callable:
         """Attach the semstring to the given object."""
@@ -1574,6 +1583,7 @@ class JacMachineInterface(
     JacBuiltin,
     JacCmd,
     JacBasics,
+    JacByLLM,
     JacUtils,
 ):
     """Jac Feature."""